Added a monitor class and associated file handling.
The monitors are callable objects.
When the monitor is called it runs the no-argument `run_filter` method.
If `run_filter` returns a truthy value then `callback` is called with the arguments given to the monitor's call.
Otherwise no action is taken.
Both routes return no value; even if the `callback` function returns a value.

Currently the two implemented monitors detect if there are files in a given folder, and if there are any files matching a given filter in a given folder.