Reworked `roles.__init__`.
Removed `RoleConfig` as all configs are passed around as `dict` and any role that needs a special configuration object can create their own class.

Changed `RoleFactory`:
 - Renamed `build` to `makeService` and it creates a new instance and configures it.
 - Created `build` to create a plain unconfigured instance of the `Role`. This is used by `makeService` to create a new instance.
 - Added a class variable to indicate the maximum number of instances that should be running.

Removed `SingletonRole` and `SingletonRoleFactory` and replaced it with an instance count defined in `RoleFactory`.

Changed `Role` to be a Service which means that implementers need to override `startService` and `stopService` for the setup and teardown code.